added env var configuration for CMS parsing errors
