+++ src/routes/+layout.svelte
@@ -15,7 +15,7 @@
   import QuoteCarousel from '$lib/components/Quote_Carousel.svelte';
   import Cta from '$lib/components/CTA.svelte';
   import CtaGroup from '$lib/components/CTA_group.svelte';
-  import * as parseModel from '$lib/js/parse_cms_models';
+  import {parseContent} from '$lib/js/parse_cms';
   import LinkButton from '../lib/components/Link_Button.svelte';
   import Icon from '../lib/components/Icon.svelte';
 
@@ -42,41 +42,6 @@
     }
   }
 
-  function parseContent(rawSections) {
-    if (!rawSections) {
-      return;
-    }
-
-    const parsedContent = [];
-    for (const rawSection of rawSections) {
-      try {
-        const section = {
-          collection: rawSection.collection,
-          props: parseModel[rawSection.collection](rawSection),
-        };
-        if (section.collection === 'heros') {
-          section.sort = rawSection.sort;
-        }
-        if (section.collection === 'contacts') {
-          section.item = {hr: rawSection.item.hr};
-        }
-        if (section.collection === 'wysiwyg') {
-          section.sort = rawSection.sort;
-        }
-        parsedContent.push(section);
-      } catch (err) {
-        console.group();
-        console.log(
-          `Error parsing ${rawSection.collection} on page ${$page_key}`,
-        );
-        console.log(err.message);
-        console.log(rawSection);
-        console.groupEnd();
-      }
-    }
-    return parsedContent;
-  }
-
   // Setting page title by retreiving translations from translations and conditionally taking
   // into account dynamic pages by using the page title attribute from the page data,
   // assigned in the dynamic pages +page.server
@@ -90,7 +55,7 @@
     $page_key === 'navbar.home' ? 'CorrelAid - Data4Good' : title_content;
 
   let content;
-  $: content = parseContent(data.builder);
+  $: content = parseContent(data.builder, $page_key);
 </script>
 
 <svelte:head>
